Update LOCATIONS_plot.R
Updated the outline for the polygons to be the same as the base color of the polygon, but with the alpha set to 1.   I had not added the call to scale_color_gradient, which allowed me to have the colors for both the points and the polygons fall into the same scale.  We have to specify the color and the fill in the aes() call.