get.overlapping bugfix
Fixing problems with looking at common ancestors; NSDUH and CBSAs both contain counties, but aren't directly associated in the source data.  So when looking for overlapping NSDUH locations for a CBSA, we look down at the counties that are contained within both the CBSAs and the NSDUH regions and match those.