Pre-merging polygons
We are moving from merging the polygons in the location.plot function to merging them in the cache step, and adding those merged locations into the location.manager as valid locations.  Introduce the PHD type, which stands for Public Health Districts,  which are combinations of counties in a given state.  We start with Alabama.